Rename iter::adapters::intersperse to ::separate
https://github.com/rust-lang/rust/issues/79524#issuecomment-2002474737
