+++ library/core/tests/iter/adapters/mod.rs
@@ -11,11 +11,11 @@ mod flat_map;
 mod flatten;
 mod fuse;
 mod inspect;
-mod intersperse;
 mod map;
 mod map_windows;
 mod peekable;
 mod scan;
+mod separate;
 mod skip;
 mod skip_while;
 mod step_by;
